chore: move ApplySingle to separated struct
