+++ pkg/template/processor_test.go
@@ -16,7 +16,6 @@ import (
 	authv1 "github.com/openshift/api/authorization/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -441,145 +440,6 @@ func TestProcessAndApply(t *testing.T) {
 				Name:       "foo",
 			},
 		}, ns.OwnerReferences)
-	})
-}
-
-func TestApplySingle(t *testing.T) {
-	// given
-	s := addToScheme(t)
-	codecFactory := serializer.NewCodecFactory(s)
-	decoder := codecFactory.UniversalDeserializer()
-
-	// create default objects with default params
-	cl := NewFakeClient(t)
-	p := template.NewProcessor(cl, s)
-	tmpl, err := DecodeTemplate(decoder,
-		CreateTemplate(WithObjects(Service, ConfigMap), WithParams(NamespaceParam, ServSelectorParam)))
-	require.NoError(t, err)
-	defaultObjs, err := p.Process(tmpl, map[string]string{})
-	require.NoError(t, err)
-	defaultObject := defaultObjs[0].Object
-
-	// create objects containing modified params
-	tmpl, err = DecodeTemplate(decoder,
-		CreateTemplate(WithObjects(Service, ConfigMap), WithParams(NamespaceParam, ServSelectorParam)))
-	require.NoError(t, err)
-	values := map[string]string{
-		"NAMESPACE":        "toolchain-host-operator",
-		"SERVICE_SELECTOR": "all-services",
-	}
-	modifiedObjs, err := p.Process(tmpl, values)
-
-	t.Run("updates of service object", func(t *testing.T) {
-		// given
-		namespacedName := types.NamespacedName{Namespace: "toolchain-host-operator", Name: "registration-service"}
-
-		t.Run("when using forceUpdate=true, it should update even when spec is same", func(t *testing.T) {
-			// given
-			cli := NewFakeClient(t)
-			processor := template.NewProcessor(cli, s)
-			_, err := processor.ApplySingle(defaultObject.DeepCopyObject(), true, nil)
-			require.NoError(t, err)
-
-			// when
-			createdOrChanged, err := processor.ApplySingle(modifiedObjs[0].Object, true, nil)
-
-			// then
-			require.NoError(t, err)
-			assert.True(t, createdOrChanged)
-		})
-
-		t.Run("when using forceUpdate=false, it should update when spec is different", func(t *testing.T) {
-			// given
-			cli := NewFakeClient(t)
-			processor := template.NewProcessor(cli, s)
-			_, err := processor.ApplySingle(defaultObject.DeepCopyObject(), true, nil)
-			require.NoError(t, err)
-
-			// when
-			createdOrChanged, err := processor.ApplySingle(modifiedObjs[0].Object, false, nil)
-
-			// then
-			require.NoError(t, err)
-			assert.True(t, createdOrChanged)
-			service := &corev1.Service{}
-			err = cli.Get(context.TODO(), namespacedName, service)
-			require.NoError(t, err)
-			assert.Equal(t, "all-services", service.Spec.Selector["run"])
-		})
-
-		t.Run("when using forceUpdate=false, it should NOT update when using same object", func(t *testing.T) {
-			// given
-			cli := NewFakeClient(t)
-			processor := template.NewProcessor(cli, s)
-			_, err := processor.ApplySingle(defaultObject.DeepCopyObject(), true, nil)
-			require.NoError(t, err)
-
-			// when
-			createdOrChanged, err := processor.ApplySingle(defaultObject, false, nil)
-
-			// then
-			require.NoError(t, err)
-			assert.False(t, createdOrChanged)
-		})
-
-		t.Run("when object is missing, it should create it no matter what is set as forceUpdate", func(t *testing.T) {
-			// given
-			cli := NewFakeClient(t)
-			processor := template.NewProcessor(cli, s)
-			require.NoError(t, err)
-			deployment := &v1.Deployment{}
-
-			// when
-			createdOrChanged, err := processor.ApplySingle(modifiedObjs[0].Object, false, deployment)
-
-			// then
-			require.NoError(t, err)
-			assert.True(t, createdOrChanged)
-			service := &corev1.Service{}
-			err = cli.Get(context.TODO(), namespacedName, service)
-			require.NoError(t, err)
-			assert.Equal(t, "all-services", service.Spec.Selector["run"])
-			assert.NotEmpty(t, service.OwnerReferences)
-		})
-
-		t.Run("when object cannot be retrieved because of any error, then it should fail", func(t *testing.T) {
-			// given
-			cli := NewFakeClient(t)
-			cli.MockGet = func(ctx context.Context, key client.ObjectKey, obj runtime.Object) error {
-				return fmt.Errorf("unable to get")
-			}
-			processor := template.NewProcessor(cli, s)
-			require.NoError(t, err)
-
-			// when
-			createdOrChanged, err := processor.ApplySingle(modifiedObjs[0].Object, false, nil)
-
-			// then
-			require.Error(t, err)
-			assert.False(t, createdOrChanged)
-			assert.Contains(t, err.Error(), "unable to get the resource")
-		})
-	})
-
-	t.Run("when using forceUpdate=false, it should update ConfigMap when data field is different", func(t *testing.T) {
-		// given
-		cli := NewFakeClient(t)
-		processor := template.NewProcessor(cli, s)
-		_, err := processor.ApplySingle(defaultObjs[1].Object, true, nil)
-		require.NoError(t, err)
-
-		// when
-		createdOrChanged, err := processor.ApplySingle(modifiedObjs[1].Object, false, nil)
-
-		// then
-		require.NoError(t, err)
-		assert.True(t, createdOrChanged)
-		configMap := &corev1.ConfigMap{}
-		namespacedName := types.NamespacedName{Namespace: "toolchain-host-operator", Name: "registration-service"}
-		err = cli.Get(context.TODO(), namespacedName, configMap)
-		require.NoError(t, err)
-		assert.Equal(t, "all-services", configMap.Data["service-selector"])
 	})
 }
